+++ xmake/toolchains/clang/xmake.lua
@@ -18,17 +18,13 @@
 -- @file        xmake.lua
 --
 
--- define toolchain
 toolchain("clang")
 
-    -- set homepage
     set_homepage("https://clang.llvm.org/")
     set_description("A C language family frontend for LLVM")
 
-    -- mark as standalone toolchain
     set_kind("standalone")
 
-    -- set toolset
     set_toolset("cc", "clang")
     set_toolset("cxx", "clang", "clang++")
     set_toolset("ld", "clang++", "clang")
@@ -40,15 +36,11 @@ toolchain("clang")
     set_toolset("mxx", "clang", "clang++")
     set_toolset("as", "clang")
 
-    -- check toolchain
     on_check(function (toolchain)
         return import("lib.detect.find_tool")("clang")
     end)
 
-    -- on load
     on_load(function (toolchain)
-
-        -- add march flags
         local march
         if toolchain:is_arch("x86_64", "x64") then
             march = "-m64"