+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [#1765](https://github.com/xmake-io/xmake/issues/1756): Support nim language
 * [#1762](https://github.com/xmake-io/xmake/issues/1762): Manage and switch the given package envs for `xrepo env`
 * [#1767](https://github.com/xmake-io/xmake/issues/1767): Support Circle compiler
+* [#1753](https://github.com/xmake-io/xmake/issues/1753): Support armcc/armclang toolchains for Keil/MDK
 
 ### Changes
 
@@ -1120,6 +1121,7 @@
 * [#1765](https://github.com/xmake-io/xmake/issues/1756): 支持 nim 语言
 * [#1762](https://github.com/xmake-io/xmake/issues/1762): 为 `xrepo env` 管理和切换指定的环境配置
 * [#1767](https://github.com/xmake-io/xmake/issues/1767): 支持 Circle 编译器
+* [#1753](https://github.com/xmake-io/xmake/issues/1753): 支持 Keil/MDK 的 armcc/armclang 工具链
 
 ### 改进